Renamed lib to library and adopted globalThis singleton pattern
