+++ demos/example-nextjs/next.config.ts
@@ -1,12 +1,9 @@
 import type { NextConfig } from 'next';
 
 const config: NextConfig = {
-  images: {
-    disableStaticImages: true
-  },
   serverExternalPackages: ['pg', 'jose'],
   turbopack: {
-    root: __dirname
+    root: import.meta.dirname
   }
 };
