+++ demos/example-nextjs/src/app/providers.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { PowerSyncProvider } from '@/lib/powersync/powersync-provider';
+import { PowerSyncProvider } from '@/library/powersync/powersync-provider';
 
 export function Providers({ children }: { children: React.ReactNode }) {
   return <PowerSyncProvider>{children}</PowerSyncProvider>;
