+++ demos/example-nextjs/src/library/auth-keys.ts
