+++ demos/example-nextjs/src/library/powersync/connector.ts
