+++ demos/example-nextjs/src/library/powersync/schema.ts
